+++ CHANGELOG.md
@@ -9,6 +9,11 @@ All notable changes to this project will be documented in this file.
 - Py dependency
 - Security issue
 
+### Features
+
+- *(config)* Introduce etcd for dynamic configuration management
+- *(helm)* Add chart version and local path fallback for pedestal deployment
+
 ### Testing
 
 - Migrate from Docker Compose to Kubernetes-based testing